Binary_Semantic_Segmentation/requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch==2.6.0+cu118
torchvision==0.21.0+cu118
torchaudio==2.6.0+cu118

--extra-index-url https://pypi.org/simple
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
wheel==0.45.1
